Complete invite participant form error handling
