+++ src/views/RecIM/components/Forms/InviteParticipantForm/index.jsx
@@ -3,26 +3,19 @@ import { useState, useEffect } from 'react';
 import GordonDialogBox from 'components/GordonDialogBox';
 import { ParticipantList } from './../../List';
 import GordonQuickSearch from 'components/Header/components/QuickSearch';
-import GordonSnackbar from 'components/Snackbar';
 import { addParticipantToTeam } from 'services/recim/team';
-import { useCallback } from 'react';
 import GordonLoader from 'components/Loader';
 import styles from './InviteParticipantForm.module.css';
 
 const InviteParticipantForm = ({
-  closeWithSnackbar,
+  createSnackbar,
   openInviteParticipantForm,
   setOpenInviteParticipantForm,
   teamID,
 }) => {
   const [disableUpdateButton, setDisableUpdateButton] = useState(true);
   const [inviteList, setInviteList] = useState([]);
-  const [snackbar, setSnackbar] = useState({ message: '', severity: null, open: false });
   const [saving, setSaving] = useState(false);
-
-  const createSnackbar = useCallback((message, severity) => {
-    setSnackbar({ message, severity, open: true });
-  }, []);
 
   useEffect(() => {
     setDisableUpdateButton(!inviteList || !inviteList.length);
@@ -44,16 +37,28 @@ const InviteParticipantForm = ({
     setInviteList(inviteList.filter((participant) => participant.Username !== username));
   };
 
-  const handleSubmit = async () => {
+  const handleSubmit = () => {
     setSaving(true);
+    let errorList = [];
     for (let index = 0; index < inviteList.length; index++) {
       let participantData = {
         Username: inviteList[index].Username,
         RoleTypeID: 2,
       };
-      await addParticipantToTeam(teamID, participantData);
+      addParticipantToTeam(teamID, participantData).catch((reason) => {
+        errorList.push({username: inviteList[index].Username, reason: reason.title})
+      });
     }
     setSaving(false);
+    let errorMessage = "Error inviting:\n"
+    errorList.forEach((value) => {
+      errorMessage += errorMessage + `${value.username} with reason ${value.reason}`
+    })
+    if (errorList.length !== 0) {
+      createSnackbar(errorMessage, 'error')
+    } else {
+      createSnackbar('Invited your participants successfully', 'success')
+    }
     handleWindowClose();
   };
 
@@ -90,12 +95,6 @@ const InviteParticipantForm = ({
             </Grid>
           )}
         </Grid>
-        <GordonSnackbar
-          open={snackbar.open}
-          text={snackbar.message}
-          severity={snackbar.severity}
-          onClose={() => setSnackbar((s) => ({ ...s, open: false }))}
-        />
         {saving && <GordonLoader size={32} />}
       </GordonDialogBox>
     </>
